Focus: fix the ring, hover/keyboard arbitration, and arrow keys (D33-D38)
Reported by the puzzle app, which had hand-rolled a replacement for each of
these. Repros land first in tests/downstream_gaps_test.cpp; each fails without
the corresponding fix.

D33 focus_ring_thickness = 0 did not disable the ring. Batched routed 0 to
render_rounded_outline_batch's else branch (1px line) and both renderers drew
the contrast outline unconditionally, so the only opt-out silently didn't work.

D34 The immediate renderer drew the ring BEFORE the widget fill, and focus_rect
is inset inside draw_rect -- so an opaque background covered it completely.
That is every default button(). Batched was correct, which is how it survived.
Extracted detail::focus_ring_for() so the two paths cannot drift again; the
ring can now sit outside the widget (focus_rect takes a float, negative offset
outsets) and draws a contrast edge on the inside as well as the outside, so it
stays visible on a widget filled with the ring colour.

D35 Hover-follow overwrote keyboard focus. ComputeVisualFocusId runs after
every user system and re-grabs focus for whatever is hot, so an app moving
focus in its build had it undone the same frame whenever the cursor rested on a
widget. Added FocusSource{Grab,Pointer,Explicit}: set_focus records intent,
hover defers to an Explicit claim. The Grab/Explicit split is load-bearing --
try_to_grab sets focus every frame, so treating that as intent would disable
hover-follow outright. Also fixes HandleFocusUICommand being undone the same
way. mouse.moved_this_frame gains a dead zone; both non-finite cases are
decided explicitly, since NaN either reported motion forever or, under a naive
distance test, poisoned prev and reported none forever.

D36 Arrows were welded to Tab, and the escape hatch was dead code -- three
references, all declarations, nothing ever attached it. Renamed to
ConsumesDirectionalInput (AcceptsValueInput kept as an alias) because a tray
stepping its children is not adjusting a value, and that framing is why nobody
wired it up. tray() now sets it, which fixes Column trays -- every dropdown's
option list -- being un-navigable. Added theme.arrows_tab so an app with its
own directional nav can opt out instead of racing for a single-slot last_action.
HandleTrayNavigation follows hot_id, so clicking a tray item no longer leaves
selection_index behind.

D37 e2e push_key used a queue with a global one-consumer-per-frame latch while
InputSystem polls each action 3x per gamepad id, so a queued key reached
whichever caller asked first and nobody else -- fine in an example, dead in a
real app. Now routed through input_injector (multi-reader); the queue stays for
push_char. The injector press lands a frame later, so the interleaving case in
input_injector_mouse_delta_test was rewritten: the property it protected is now
true by construction.

Verified against wm_afterhours at its pinned 14222a3 with this patch applied:
tray_vertical_navigation FAIL -> PASS (Down was tabbing out of the vertical
tray), and validate-screenshots is unaffected -- the same 8 pre-existing
failures before and after, three moving by <=0.0014 percentage points.

+++ src/plugins/ui/ui_core_components.h
@@ -181,9 +181,12 @@ struct UIComponent : BaseComponent {
   float width() const { return rect().width; }
   float height() const { return rect().height; }
 
-  Rectangle focus_rect(int rw = 4) const {
-    return Rectangle{x() + (float)rw, y() + (float)rw,
-                     width() - (2.f * (float)rw), height() - (2.f * (float)rw)};
+  // Positive `rw` insets, negative outsets. Float, not int: the theme's
+  // focus_ring_offset is a float, and truncating it here made sub-pixel
+  // offsets unreachable.
+  Rectangle focus_rect(float rw = 4.f) const {
+    return Rectangle{x() + rw, y() + rw, width() - 2.f * rw,
+                     height() - 2.f * rw};
   }
 
   auto &make_absolute() {
